services/mss_chuanyin/pyproject.toml
[project]
name = "mss-chuanyin"
version = "0.1.0"
description = "miaosuan service chuanyin"
authors = [
    {name = "kevincnzhengyang",email = "kevin.cn.zhengyang@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.13.5"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "line-bot-sdk (>=3.18.1,<4.0.0)",
    "click (>=8.2.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
